refactor(server): split mobility preflight service
