+++ server/proliferate/server/cloud/mobility/models.py
@@ -85,27 +85,6 @@ class EnsureMobilityWorkspaceRequest(BaseModel):
     owner_hint: str = Field(default="local", alias="ownerHint")
 
 
-class WorkspaceMobilityPreflightRequest(BaseModel):
-    direction: str
-    requested_branch: str = Field(alias="requestedBranch")
-    requested_base_sha: str | None = Field(default=None, alias="requestedBaseSha")
-
-
-class WorkspaceMobilityPreflightBlocker(BaseModel):
-    code: str
-    message: str
-    source: str = "cloud"
-    retry_action: str | None = Field(default=None, serialization_alias="retryAction")
-    details: dict[str, str] | None = None
-
-
-class WorkspaceMobilityPreflightResponse(BaseModel):
-    can_start: bool = Field(serialization_alias="canStart")
-    blockers: list[WorkspaceMobilityPreflightBlocker]
-    excluded_paths: list[str] = Field(serialization_alias="excludedPaths")
-    workspace: MobilityWorkspaceDetail
-
-
 class StartWorkspaceMobilityHandoffRequest(BaseModel):
     direction: str
     requested_branch: str = Field(alias="requestedBranch")
