+++ server/proliferate/server/cloud/mobility/preflight/__init__.py
@@ -1,0 +1,1 @@
+"""Workspace mobility preflight subdomain."""
